Afisha-ArsenHomework--main/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
pip-system-certs==4.0
sqlparse==0.5.2
typing_extensions==4.12.2
wrapt==1.17.0
